% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guiMCP.R
\name{guiMCP}
\alias{guiMCP}
\title{Graphical User Interface for MCPtests package}
\usage{
guiMCP(gui = TRUE)
}
\arguments{
\item{gui}{Logical argument, \code{TRUE} or \code{FALSE}. The default is \code{TRUE}}
}
\value{
\code{guiMCP} presents a GUI for the results of the multiple
      comparison procedures in the literature. In addition, the GUI
      returns a graph of the results, as well as the export of these results
      to three types of file extension and latex code.
}
\description{
\code{guiMCP} A Graphical User Interface (GUI) for
    the MCP package
}
\examples{
# Loading package
library(MCPtests)
if (interactive()) {
  guiMCP(gui = FALSE)
}

}
